Broke enums away from merc.h
